Adjusments according to logic and server recent changes related to record model. - Model properties updated in forms. - User-friendly formats applied: date and currency. - Date management adjustments to match server / database format. - API service adjustments for server's endpoints. - State data persistence re-configured to save required data (localstorage. To change in near future). - Fixed to fund model management: balance property management in forms and cards, to display and calculate according to records timeline.

+++ src/services/fundAPI.js
@@ -1,5 +1,5 @@
 import API from './API';
 
-export const Create = (body) => API.post('/funds', body);
-export const Update = ({ id, body }) => API.patch(`/funds/${id}`, body);
-export const Delete = (id) => API.delete(`/funds/${id}`);
+export const Create = ({ token, ...body }) => API.post('/funds', body, { headers: { Authorization: `bearer ${token}` } });
+export const Update = ({ id, body, token  }) => API.patch(`/funds/${id}`, body, { headers: { Authorization: `bearer ${token}` } });
+export const Delete = ({ id, token }) => API.delete(`/funds/${id}`, { headers: { Authorization: `bearer ${token}` } });
